requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
tensorflow==2.17.0
pillow
numpy
python-multipart
